Remove note component from image component
