+++ components/blocks/note.js
@@ -1,39 +1,24 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
+import classNames from "classnames";
 
-const Note = ({ dark, background, color, children }) => {
-  const [theme, setTheme] = useState("light-mode");
+import CalloutStyles from "./callout.module.css";
+import NoteStyles from "./note.module.css";
 
-  useEffect(() => {
-    window.addEventListener("ChangeTheme", handleTheme);
+import IconHeader from "./iconHeader";
 
-    return () => {
-      window.removeEventListener("ChangeTheme", handleTheme);
-    };
-  }, []);
-
-  const handleTheme = () => {
-    setTheme(document.body.dataset.theme);
-  };
-
-  let block;
-
-  if (dark && theme == "dark-mode") {
-    block = (
-      <section
-        className={`block-note bg-${dark.background} color-${dark.color}`}
-      >
-        {children}
-      </section>
-    );
-  } else {
-    block = (
-      <section className={`block-note bg-${background} color-${color}`}>
-        {children}
-      </section>
-    );
-  }
-
-  return block;
+const Note = ({ children }) => {
+  return (
+    <section className={classNames(CalloutStyles.Container, NoteStyles.Note)}>
+      <IconHeader
+        icon="push_pin"
+        rotate="45"
+        title="Note"
+        background="l-blue-70"
+        color="white"
+      />
+      {children}
+    </section>
+  );
 };
 
 export default Note;
